Added Animation and Plot as separate options
inverted_pendulum_animated.py:
Created a version of the simulation that uses pygame to animate the motion of the pendulum in a continuous and ongoing manner - until the user closes the window.

inverted_pendulum_graphed.py
Separated that from another file that calculates the behavior of the pendulum over a specified number of time steps, and then produces static graphs at the end.