Found a sample case
The hard-coded sample case (FAx_0 = 5, w = 3.2, theta_0 = 75 deg) demonstrates the basic control idea. From this angle, the bar should fall CW. Instead, the oscillating forcing function applied to the base of the pendulum causes it to rotate CCW. It balances for a few oscillations before falling CCW - because the control must actively modify these parameters to truly balance the rod.